pyproject.toml
[tool.poetry]
name = "mcp-memory-server"
version = "0.1.0"
description = "MCP Server with Qdrant Vector Database Memory System for Cursor"
authors = ["Hannes <hannes@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10,<3.13"
mcp = "^1.13.1"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.5.1"
numpy = "^1.26.0"
python-dotenv = "^1.0.0"
markdown = "^3.5.0"
beautifulsoup4 = "^4.12.0"
aiofiles = "^24.1.0"
pyyaml = "^6.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.17.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
types-markdown = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"